harness/requirements.txt
kernel>=0.52.0
tzafon>=0.1.0
google-genai>=1.51.0
openai>=1.50.0
python-dotenv>=1.0.0
playwright>=1.49.0
httpx>=0.27.0
pillow>=10.0.0
